Phone: Incoming Call in Background
Allows user to set that Incoming Call UI will remain in background
if screen is on. Instead, a notification is added to StatusBar, which
can be used to bring up the Incoming Call UI. Ringtone and vibration
still occur as per system settings.

This has the same effect as apps marked as "immersive". It prevents
the incoming call from backgrounding the current activity.

PS 2: minor fix

Change-Id: Ib98dd74268826d511ce43ac75dd9dd66f774c9e7